+++ src/main/java/com/ll/medium/domain/member/member/entity/Member.java
@@ -1,6 +1,6 @@
 package com.ll.medium.domain.member.member.entity;
 
-import com.ll.medium.global.jpa.baseEntity.BaseEntity;
+import com.ll.medium.global.jpa.BaseEntity;
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
 import lombok.AllArgsConstructor;
